+++ zcash_primitives/CHANGELOG.md
@@ -82,6 +82,9 @@ and this library adheres to Rust's notion of
   - `impl From<NonNegativeAmount> for zcash_primitives::sapling::value::NoteValue`
   - `impl Sum<NonNegativeAmount> for Option<NonNegativeAmount>`
   - `impl<'a> Sum<&'a NonNegativeAmount> for Option<NonNegativeAmount>`
+- `impl {Clone, PartialEq, Eq} for zcash_primitives::memo::Error`
+- `impl {PartialEq, Eq} for zcash_primitives::sapling::note::Rseed`
+- `impl From<TxId> for [u8; 32]`
 
 ### Changed
 - `zcash_primitives::sapling`: